+++ backend/ddang/src/main/java/com/ddang/ddang/review/presentation/ReviewController.java
@@ -5,8 +5,8 @@ import com.ddang.ddang.authentication.domain.dto.AuthenticationUserInfo;
 import com.ddang.ddang.review.application.ReviewService;
 import com.ddang.ddang.review.application.dto.CreateReviewDto;
 import com.ddang.ddang.review.application.dto.ReadReviewDto;
-import com.ddang.ddang.review.presentation.request.CreateReviewRequest;
-import com.ddang.ddang.review.presentation.response.ReadReviewResponse;
+import com.ddang.ddang.review.presentation.dto.request.CreateReviewRequest;
+import com.ddang.ddang.review.presentation.dto.response.ReadReviewResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
